refactor AbstractManager and AbstractConnectionManager
